requirements.txt
bcrypt==4.0.1
cachetools==5.3.3
cryptography==41.0.1
email-validator==1.3.1
fastapi==0.81.0
passlib==1.7.4
pydantic==1.10.1
PyMySQL==1.0.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
sqlalchemy==1.4.40
uvicorn==0.18.3
